Resolving FxCop and StyleCop warnings
Updating to conform to FxCop and StyleCop standards. The StyleCop rules
used are the default, including documenting private members and using
clauses inside namespaces (both of which I usually turn off by default).

The FxCop rule set used in each project is the "Microsoft Extended Design
Guideline Rules." This is used to "enforce the highest standards for
writing maintainable code." Suppression of specific warnings is used for
methods taking strings for URI parameters instead of System.Uri for ease
of use, and for unused parameters in the Main() method of the console
applications.

Copyright headers have been added to each file, along with licensing
information. The license is basically the MIT license; do with this code
whatever you please.